Split the insight.py implementation into two distinct versions
Too complex to share the logic between `./backend` and `./cli`, and in
the end the tests were overlapping since we expected a file for the CLI
containing the opt_out answer from the user, whereas we expect an
environment variable for the backend logic.
